Flowchart Refactoring
Creating new slice for flowchart selection state vars to seperate concerns from flowchart data information
