+++ Client/src/redux/index.ts
@@ -10,7 +10,8 @@ export { layoutReducer } from "./layout/layoutSlice.ts";
 export { logReducer } from "./log/logSlice.ts";
 export { authReducer } from "./auth/authSlice.ts";
 export { userReducer } from "./user/userSlice.ts";
-
+export { flowchartReducer } from "./flowchart/flowchartSlice.ts";
+export { flowSelectionReducer } from "./flowSelection/flowSelectionSlice.ts";
 // Export all actions for easy dispatching in components
 export * as messageActions from "./message/messageSlice.ts";
 export * as gptActions from "./gpt/gptSlice";
@@ -19,6 +20,7 @@ export * as logActions from "./log/logSlice.ts";
 export * as authActions from "./auth/authSlice.ts";
 export * as userActions from "./user/userSlice.ts";
 export * as flowchartActions from "./flowchart/flowchartSlice.ts";
+export * as flowSelectionActions from "./flowSelection/flowSelectionSlice.ts";
 // Custom hooks for using dispatch and selector with TypeScript
 import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
 import { RootState, AppDispatch } from "./store";
